refactored question-details by extracting flattenAnswers helper function
